Double the number of parts when computing the closest point to the contiuous line
I found cases for slabs which are really large and only constrained by a
few points at which the closest point is very wrong at some locations.
It is an unlikelysenario, but I want to have it covered by default. I
think I should make it an input value at some point.